1) Added "Top 8" on top of karts icons on RaceGUI when there are more than 8 karts displayed on RaceGUI.
2) Fixed messages of RaceGUI (by giving them lower sizes).
3) Added one step to the next sector in handleSteering (to fix karts crashing in other karts having a plunger, because they are too slow).
4) Cosmetic changes.

+++ src/gui/font.cpp
@@ -94,7 +94,7 @@ void Font::Print(const char *text, float size, float x, float y,
     }
 
     m_text_out->begin();
-    m_text_out->setPointSize((float)sz);
+    m_text_out->setPointSize(sz);
     if(doShadow)
     {
         m_text_out->start2f(x-2, y-2);
